Add X-Request-ID and X-Correlation-ID headers
